frontend: rename request related visual components
These components to relate in future to a transfer and thereby follow
the new terminology.

As the affects by the changes to all Vue components will become quite
massive due the additional renaming, test changes etc. pp., this small
commit helps to split this change into multiple steps. This allows to
keep easier track of what actually happens.
