+++ frontend/tests/unit/components/TransferStatus.spec.ts
@@ -1,13 +1,13 @@
 import { mount } from '@vue/test-utils';
 
-import RequestProcessing from '@/components/RequestProcessing.vue';
-import RequestStatus from '@/components/RequestStatus.vue';
-import RequestSummary from '@/components/RequestSummary.vue';
+import TransferProcessing from '@/components/TransferProcessing.vue';
+import TransferStatus from '@/components/TransferStatus.vue';
+import TransferSummary from '@/components/TransferSummary.vue';
 import { RequestMetadata, RequestState } from '@/types/data';
 import { generateRequestMetadata } from '~/utils/data_generators';
 
 function createWrapper(options?: { metadata?: RequestMetadata; state?: RequestState }) {
-  return mount(RequestStatus, {
+  return mount(TransferStatus, {
     shallow: true,
     props: {
       metadata: options?.metadata ?? generateRequestMetadata(),
@@ -23,8 +23,8 @@ function createWrapper(options?: { metadata?: RequestMetadata; state?: RequestSt
   });
 }
 
-describe('RequestStatus.vue', () => {
-  it('shows request summary with correct metadata', () => {
+describe('TransferStatus.vue', () => {
+  it('shows transfer summary with correct metadata', () => {
     const metadata = generateRequestMetadata({
       amount: '0.1',
       tokenSymbol: 'TTT',
@@ -33,7 +33,7 @@ describe('RequestStatus.vue', () => {
       targetAddress: '0xAddress',
     });
     const wrapper = createWrapper({ metadata });
-    const summary = wrapper.findComponent(RequestSummary);
+    const summary = wrapper.findComponent(TransferSummary);
 
     expect(summary.exists()).toBeTruthy();
     expect(summary.isVisible()).toBeTruthy();
@@ -44,9 +44,9 @@ describe('RequestStatus.vue', () => {
     expect(summary.props()).toContain({ targetAddress: '0xAddress' });
   });
 
-  it('shows request processing with correct state', () => {
+  it('shows transfer processing with correct state', () => {
     const wrapper = createWrapper({ state: RequestState.WaitFulfill });
-    const process = wrapper.findComponent(RequestProcessing);
+    const process = wrapper.findComponent(TransferProcessing);
 
     expect(process.exists()).toBeTruthy();
     expect(process.isVisible()).toBeTruthy();
